+++ PAF/ReleaseNotes.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - search functionality
 - save configuration settings
 - `Write-ErrorLog`
+- inline help for functions
 
 ### Changed